+++ internal/gridapi/project.go
@@ -177,6 +177,7 @@ func projectNetwork(ev *gridv1.Event) hazards.Feature {
 		SNR:       t.GetSnr(),
 		RSSI:      t.GetRssi(),
 		HopCount:  t.GetHopCount(),
+		Path:      t.GetPath(),
 		Gateways:  t.GetGateways(),
 	}
 	return feature(ev, p)